Handle scroll down
-Handle scroll down
-Delete the above data

+++ var.go
@@ -8,7 +8,7 @@ import (
 
 var (
 	CurrentDBClient dbpak.DBClient
-	CurrentPage     uint8
+	CurrentPage     int
 	ItemsPerPage    = 20
 	FolderPath      string
 	CurrentJson     jsFile.JsonFile
